fix: move dialog outside fetch posts component
